TASK1
This Online Reservation System will include all the necessary fields which are required during
online reservation system. This Online Reservation System will be easy to use and can be used by
any person. The basic idea behind this project is to save data in a central database which can be
accessed by any authorize person to get information and saves time and burden which are being
faced by their customers.